Method addion and refactoring
GetEvents method added, which returns an IEnumerable object of all the
events in this collection. A filter can be applied to return certain
types of events.

ToCoordinateList() method now returns a CoordianteCollection rather than
a List<Coordinate>

A clear all clustering method has been added to remove any meta-data
that is used by the clustering algorithms.